separate cluster and vng types into different files
